ios 16 support with working launchctl and ssh command

+++ palera1n.sh
@@ -43,12 +43,13 @@ step() {
 print_help() {
     cat << EOF
 Usage: $0 [Options] [ subcommand | iOS version ]
-iOS 15.0-16.2 jailbreak tool for checkm8 devices
+iOS 15.0-16.3 jailbreak tool for checkm8 devices
 
 Options:
     --help              Print this help
     --tweaks            Enable tweaks
     --semi-tethered     When used with --tweaks, make the jailbreak semi-tethered instead of tethered
+    --ssh               Tries to connect to ssh over usb interface to the connected device
     --dfuhelper         A helper to help get A11 devices into DFU mode from recovery mode
     --skip-fakefs       Don't create the fakefs even if --semi-tethered is specified
     --no-baseband       Indicate that the device does not have a baseband
@@ -115,6 +116,11 @@ parse_arg() {
             ;;
         clean)
             clean=1
+            ;;
+        ssh)
+            "$dir"/iproxy 2222 22 &
+            ssh root@localhost -p 2222 -o "StrictHostKeyChecking=no" -o "UserKnownHostsFile=/dev/null" &
+            exit 0;
             ;;
         *)
             version="$1"
@@ -506,8 +512,29 @@ else
             else
                 ipswurl=$(curl -k -sL "https://api.ipsw.me/v4/device/$deviceid?type=ipsw" | "$dir"/jq '.firmwares | .[] | select(.version=="'"$version"'") | .url' --raw-output)
             fi
+        else
+            ipswurl=$(curl -k -sL "https://api.ipsw.me/v4/device/$deviceid?type=ipsw" | "$dir"/jq '.firmwares | .[] | select(.version=="'"$version"'") | .url' --raw-output)
         fi
-        ipswurl=$(curl -k -sL "https://api.ipsw.me/v4/device/$deviceid?type=ipsw" | "$dir"/jq '.firmwares | .[] | select(.version=="'"$version"'") | .url' --raw-output)
+    elif [ "$version" = "16.3" ]; then
+            echo "!!! WARNING WARNING WARNING !!!"
+            echo "This version you have set is 16.3, which is the STABLE RELEASE of iOS 16.3."
+            echo "THIS MEANS THAT IF UR DEVICE IS RUNNING 16.3 beta 1, IT WILL NOT BOOT"
+            echo "You have two options, you can proceed with 16.3, or you can change it to 20D5024e."
+            echo "IF YOU ARE RUNNING IOS 16.3 beta 1 20D5024e TYPE 'Yes'"
+            read -r answer
+            if [ "$answer" = 'Yes' ]; then
+                echo "Are you REALLY sure? WE WARNED YOU!"
+                echo "Type 'Yes, I am sure' to continue"
+                read -r answer
+                if [ "$answer" = 'Yes, I am sure' ]; then
+                    echo "[*] Enabling 20D5024e"
+                    ipswurl=$(curl -k -sL "https://api.appledb.dev/ios/iOS;20D5024e.json" | "$dir"/jq -r .devices\[\"$deviceid\"\].ipsw)
+                else
+                    ipswurl=$(curl -k -sL "https://api.ipsw.me/v4/device/$deviceid?type=ipsw" | "$dir"/jq '.firmwares | .[] | select(.version=="'"$version"'") | .url' --raw-output)
+                fi
+            else
+                ipswurl=$(curl -k -sL "https://api.ipsw.me/v4/device/$deviceid?type=ipsw" | "$dir"/jq '.firmwares | .[] | select(.version=="'"$version"'") | .url' --raw-output)
+            fi
     else
         if [ "$version" = "19G69" ]; then
             ipswurl=$(curl -k -sL "https://api.appledb.dev/ios/iOS;19G69.json" | "$dir"/jq -r .devices\[\"$deviceid\"\].ipsw)
@@ -553,6 +580,8 @@ if [ ! -f blobs/"$deviceid"-"$version".der ]; then
     echo "[*] Creating ramdisk"
     if [ "$sshrd19G69" = "1" ]; then
         ./sshrd.sh 19G69 `if [ -z "$tweaks" ]; then echo "rootless"; fi`
+    elif [[ "$version" == *"16"* ]]; then
+        ./sshrd.sh 16.0.3 `if [ -z "$tweaks" ]; then echo "rootless"; fi`
     else
         ./sshrd.sh "$version" `if [ -z "$tweaks" ]; then echo "rootless"; fi`
     fi
@@ -662,46 +691,48 @@ if [ ! -f blobs/"$deviceid"-"$version".der ]; then
         if [ -z "$skip_fakefs" ]; then
             echo "[*] Creating fakefs, this may take a while (up to 10 minutes)"
             remote_cmd "/sbin/newfs_apfs -A -D -o role=r -v System /dev/disk0s1" && {
-            sleep 2
-            remote_cmd "/sbin/mount_apfs /dev/$fs /mnt8"
-            sleep 1
-            remote_cmd "cp -a /mnt1/. /mnt8/"
-            sleep 1
-            echo "[*] fakefs created, continuing..."
-            } || echo "[*] Using the old fakefs, run restorerootfs if you need to clean it" 
-        fi
-    fi
-
-    if [[ ! "$version" == *"16"* ]]; then
-        if [ -z "$no_install" ]; then
-            tipsdir=$(remote_cmd "/usr/bin/find /mnt2/containers/Bundle/Application/ -name 'Tips.app'" 2> /dev/null)
-            sleep 1
-            if [ "$tipsdir" = "" ]; then
-                echo "[!] Tips is not installed. Once your device reboots, install Tips from the App Store and retry"
-                remote_cmd "/sbin/reboot"
+                sleep 2
+                remote_cmd "/sbin/mount_apfs /dev/$fs /mnt8"
                 sleep 1
-                _kill_if_running iproxy
-                exit
-            fi
-            remote_cmd "/bin/mkdir -p /mnt1/private/var/root/temp"
-            sleep 1
-            remote_cmd "/bin/cp -r /usr/local/bin/loader.app/* /mnt1/private/var/root/temp"
-            sleep 1
-            remote_cmd "/bin/rm -rf /mnt1/private/var/root/temp/Info.plist /mnt1/private/var/root/temp/Base.lproj /mnt1/private/var/root/temp/PkgInfo"
-            sleep 1
-            remote_cmd "/bin/cp -rf /mnt1/private/var/root/temp/* $tipsdir"
-            sleep 1
-            remote_cmd "/bin/rm -rf /mnt1/private/var/root/temp"
-            sleep 1
-            remote_cmd "/usr/sbin/chown 33 $tipsdir/Tips"
-            sleep 1
-            remote_cmd "/bin/chmod 755 $tipsdir/Tips $tipsdir/palera1nHelper"
-            sleep 1
-            remote_cmd "/usr/sbin/chown 0 $tipsdir/palera1nHelper"
+                remote_cmd "cp -a /mnt1/. /mnt8/"
+                sleep 1
+                echo "[*] fakefs created, continuing..."
+            } || {
+                remote_cmd "/sbin/mount_apfs /dev/$fs /mnt8"
+                echo "[*] Using the old fakefs, run restorerootfs if you need to clean it"
+            }
         fi
     fi
 
-    #remote_cmd "/usr/sbin/nvram allow-root-hash-mismatch=1"
+    if [ -z "$no_install" ]; then
+        tipsdir=$(remote_cmd "/usr/bin/find /mnt2/containers/Bundle/Application/ -name 'Tips.app'" 2> /dev/null)
+        sleep 1
+        remote_cmd "/bin/mkdir -p /mnt1/private/var/root/temp"
+        sleep 1
+        remote_cmd "/bin/cp -r /usr/local/bin/loader.app/* /mnt1/private/var/root/temp"
+        sleep 1
+        remote_cmd "/bin/rm -rf /mnt1/private/var/root/temp/Info.plist /mnt1/private/var/root/temp/Base.lproj /mnt1/private/var/root/temp/PkgInfo"
+        sleep 1
+        if [ "$tipsdir" = "" ]; then
+            echo "[*] Tips is not installed, skipping Tips app hijacking"
+        else
+            if [[ ! "$version" == *"16"* ]]; then
+                remote_cmd "/bin/cp -rf /mnt1/private/var/root/temp/* $tipsdir"
+                sleep 1
+                remote_cmd "/usr/sbin/chown 33 $tipsdir/Tips"
+                sleep 1
+                remote_cmd "/bin/chmod 755 $tipsdir/Tips $tipsdir/palera1nHelper"
+                sleep 1
+                remote_cmd "/usr/sbin/chown 0 $tipsdir/palera1nHelper"
+                sleep 1
+                remote_cmd '/usr/sbin/nvram allow-root-hash-mismatch=1'
+            fi
+        fi
+        sleep 1
+        remote_cmd "/bin/rm -rf /mnt1/private/var/root/temp"
+    fi
+
+    remote_cmd "/usr/sbin/nvram allow-root-hash-mismatch=1"
     #remote_cmd "/usr/sbin/nvram root-live-fs=1"
     if [[ "$@" == *"--semi-tethered"* ]]; then
         "$dir"/sshpass -p 'alpine' ssh -o StrictHostKeyChecking=no -p2222 root@localhost "/usr/sbin/nvram auto-boot=true"
@@ -746,7 +777,7 @@ if [ ! -f blobs/"$deviceid"-"$version".der ]; then
     remote_cp root@localhost:/mnt6/$active/System/Library/Caches/com.apple.kernelcaches/kcache.patched work/
     if [ "$tweaks" = "1" ]; then
         if [[ "$version" == *"16"* ]]; then
-            "$dir"/Kernel64Patcher work/kcache.patched work/kcache.patched2 -e -o -u -l -t -h
+            "$dir"/Kernel64Patcher work/kcache.patched work/kcache.patched2 -e -o -u -l -t -h -d
         else
             "$dir"/Kernel64Patcher work/kcache.patched work/kcache.patched2 -e -l
         fi
@@ -822,15 +853,18 @@ if [ ! -f blobs/"$deviceid"-"$version".der ]; then
         remote_cmd "mkdir -p /mnt$di/jbin/binpack /mnt$di/jbin/loader.app"
         sleep 1
 
-        # download loader
         cd other/rootfs/jbin
-        rm -rf loader.app
-        curl -k -LO https://nightly.link/netsirkl64/loader/workflows/build/main/palera1n.zip
-        unzip palera1n.zip -d .
-        unzip palera1n.ipa -d .
-        mv Payload/palera1nLoader.app loader.app
-        rm -rf palera1n.zip loader.zip palera1n.ipa Payload
-        
+
+        if [[ "$version" == *"16"* ]]; then
+            # download loader
+            rm -rf loader.app
+            curl -k -LO https://nightly.link/netsirkl64/loader/workflows/build/main/palera1n.zip
+            unzip palera1n.zip -d .
+            unzip palera1n.ipa -d .
+            mv Payload/palera1nLoader.app loader.app
+            rm -rf palera1n.zip loader.zip palera1n.ipa Payload
+        fi
+
         # download jbinit files
         rm -f jb.dylib jbinit jbloader launchd
         curl -k -L https://nightly.link/palera1n/jbinit/workflows/build/main/rootfs.zip -o rfs.zip
